Starting to grasp Ruby - add some of the objects
The Python version of the game class initializes its data in some
concise lines, mostly dict comprehensions.

I'm starting to wrap my head around code blocks and the best way to use
them, and as a result have managed to get a pretty lean init in the Ruby
version of that class.