Started including tests; removed turn for now
The test framework is set up; it's a mixin module similar to the Python
version. A couple of basic tests added; enough to parse the first turn
of Starflight.

Removed the turn class for now; it basically just contained almost all
the methods from the adventure class. Might move it back out in the
future. Mainly just want to be able to store commands/actions so we can
replay them to restore saved games.